* review_token:
  신규 로그인 및 로그아웃 기능 적용
  레시피 상세 페이지 내 jwt 토큰 세션 에러 수정
  jwt 토큰을 이용한 레시피 상세 및 리뷰 기능 구현

+++ route/review.py
@@ -4,7 +4,6 @@ import requests, jwt
 from bs4 import BeautifulSoup
 from pymongo import MongoClient
 from bson.objectid import ObjectId
-# client = MongoClient('mongodb+srv://sparta:test@cluster0.cirioky.mongodb.net/?retryWrites=true&w=majority')
 client = MongoClient('mongodb+srv://sparta:test@cluster0.er12y5s.mongodb.net/?retryWrites=true&w=majority')
 db = client.dbsparta
 SECRET_KEY = 'recipe'
@@ -18,18 +17,22 @@ def home():
 # 리뷰 등록
 @blue_review.route("/", methods=["POST"])
 def add():
-    recipe_id_receive = request.form['recipe_id']
-    id_receive = request.form['id']
-    review_receive = request.form['review']
+    token_receive = request.cookies.get('mytoken')
+    if token_receive is not None: # token 값이 있다면 (로그인 상태라면)
+        recipe_id_receive = request.form['recipe_id']
+        id_receive = request.form['id']
+        review_receive = request.form['review']
     
-    doc = {
-        'recipe_id':recipe_id_receive,
-        'id':id_receive,
-        'review':review_receive
-        }
-    db.review.insert_one(doc)
-    
-    return jsonify({'msg':'저장완료!'})
+        doc = {
+            'recipe_id':recipe_id_receive,
+            'id':id_receive,
+            'review':review_receive
+            }
+        db.review.insert_one(doc)
+        
+        return jsonify({'msg':'저장 완료!', 'login':'true'})
+    else: # token 값이 없다면 (로그아웃 상태라면)
+        return jsonify({'msg':'먼저 로그인을 해주세요!', 'login':'false'})
 
 # 리뷰 수정
 @blue_review.route("/", methods=["PUT"])
@@ -38,14 +41,25 @@ def update():
     id_receive = request.form['id']
     review_receive = request.form['review']
 
-    # token_receive = request.cookies.get('mytoken')
-    # try:
-    #      payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
-    #      print("------------------------"+payload)
-    # except jwt.ExpiredSignatureError:
-    #     return redirect("http://localhost:5000/recipe")
-    # except jwt.exceptions.DecodeError:
-    #     return redirect("http://localhost:5000/recipe")
+    token_receive = request.cookies.get('mytoken')
+    try:
+        if token_receive is not None: # token 값이 있다면 (로그인 상태라면)
+            payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+            user_id = payload['id']
+            if id_receive == user_id: # 로그인 사용자와 해당 리뷰를 남긴 사용자가 같으면
+                doc = {
+                    'review':review_receive
+                }
+                db.review.update_one({'_id':ObjectId(under_id_receive)}, {'$set': doc})
+                return jsonify({'msg':'수정 완료!', 'login':'true'})
+            else: # 다르면
+                return jsonify({'msg':'본인이 작성한 리뷰만 수정 가능합니다!', 'login':'true'})
+        else: # token 값이 없다면 (로그아웃 상태라면)
+            return jsonify({'msg':'먼저 로그인을 해주세요!', 'login':'false'})
+    except jwt.ExpiredSignatureError:
+        return redirect("/recipe")
+    except jwt.exceptions.DecodeError:
+        return redirect("/recipe")
 
     # if id_receive == session_id:
     #     doc = {
@@ -73,11 +87,11 @@ def update():
     #     print("로그인 정보 잘못됨")
     #     return redirect("/user/signin")
 
-    doc = {
-            'review':review_receive
-    }
-    db.review.update_one({'_id':ObjectId(under_id_receive)}, {'$set': doc})
-    return jsonify({'msg':'수정완료!'})
+    # doc = {
+    #         'review':review_receive
+    # }
+    # db.review.update_one({'_id':ObjectId(under_id_receive)}, {'$set': doc})
+    # return jsonify({'msg':'수정완료!'})
     
 
 # 리뷰 삭제
@@ -86,13 +100,19 @@ def review_delete():
     id_receive = request.form['id_give'] ## 댓글의 id
     under_id_receive = request.form['under_id_give']
 
-    # print(session_id)
-    # if id_receive == session_id:
-    #     db.review.delete_one({'_id':ObjectId(under_id_receive)})
-    #     return jsonify({'msg':'삭제완료!'})
-    # else:
-    #     return jsonify({'msg':'본인의 댓글만 삭제 가능합니다!'})
-    
-    db.review.delete_one({'_id':ObjectId(under_id_receive)})
-    return jsonify({'msg':'삭제완료!'})
-    
+    token_receive = request.cookies.get('mytoken')
+    try:
+        if token_receive is not None: # token 값이 있다면 (로그인 상태라면)
+            payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+            user_id = payload['id']
+            if id_receive == user_id: # 로그인 사용자와 해당 리뷰를 남긴 사용자가 같으면
+                db.review.delete_one({'_id':ObjectId(under_id_receive)})
+                return jsonify({'msg':'삭제 완료!', 'login':'true'})
+            else: # 다르면
+                return jsonify({'msg':'본인이 작성한 리뷰만 삭제 가능합니다!', 'login':'true'})
+        else: # token 값이 없다면 (로그아웃 상태라면)
+            return jsonify({'msg':'먼저 로그인을 해주세요!', 'login':'false'})
+    except jwt.ExpiredSignatureError:
+        return redirect("/recipe")
+    except jwt.exceptions.DecodeError:
+        return redirect("/recipe")
